fix: dune describe & virtual libraries
Make `dune describe` correctly handle overlapping implementations for
virtual libraries.

Fixes issue #5747

Here is what happened before this fix: if your project was using a
virtual library and at least two implementations for that virtual
library, then dune describe would not produce any information for the
libraries of your project (it would only show information about the
executables, if there were any). The issue was in the computation of the
transitive closure of libraires, that raised an error by default in the
presence of overlapping implementations of virtual libraries. In such a
case, the error was treated by discarding all information about the
libraries.

The fix consists in computing the transitive closure of libraries
without performing any overlap checks between implementations of a
virtual library. We have defined the `descriptive_closure` function for
this very purpose.

The overlap check is necessary when you compute the transitive closure
of libraries that you intend to build (otherwise you do not know which
implementation should be picked).

In dune describe, you want to describe the whole workspace, regardless
of what you want to build or not, which means that you need to handle
all the libraries of your workspace, that may include several
implementations for virtual modules.

If you absolutely want to perform the overlap check in dune describe,
then will not always be able to describe the whole workspace.

Signed-off-by: Benoît Montagu <benoit.montagu@inria.fr>