feature: dune describe workspace accepts dirs
allow `dune describe workspace` to accept directories as arguments

The provided directories restrict the worskpace description to those
directories. Note that the transitive dependencies of the discovered
executables and libraries are always printed.

If no argument is provided, or if `.` is provided, then the behavior
is unchanged.

This fixes issue#3893.

Signed-off-by: Benoît Montagu <benoit.montagu@inria.fr>