Disable dune cache for test suite
The test suite does not behave correctly when `DUNE_CACHE=enabled` is
present in the environment. So we unset the variable locally for tests
that do not pass.

This can be tricky to test. One way that works is by exporting
`DUNE_CACHE=enabled` and running `./dune.exe runtest -f`.

Signed-off-by: Etienne Millon <me@emillon.org>