test: Move test for github issue 4345
The issue was a dependency cycle when a public library copies files from
a parent directory using copy_files. The test was that a readible error
message is produced in that situation, but it's not necessary that this
situation result in a dependency cycle at all.

This change moves the test case to its own file and now asserts that the
case doesn't result in an error.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>