feature(coq): add rules for coqffi
This adds a coq.ffi stanza with a (modules ) field which turns
specified .mli files into .v files using coqffi.

The intended use case is for the .mli file to become the interface for
the .ml file extracted from the .v file.

TODO: Add fields for more arguments to coqffi?

Signed-off-by: Ali Caglayan <alizter@gmail.com>

ps-id: 0347ad6c-3758-4184-8bcf-29fdee6b1f29