feature: Support CLICOLOR and CLICOLOR_FORCE to manage ANSI colors
The patch follows the spec at <https://bixense.com/clicolors/> that
has been implemented by multiple build systems and compilers already.

When Dune is not run in a tty (such as a CI system or a script), it
will strip ANSI colors from the output of targets (such a tests
runs). This is often undesirable.

Dune will now recognize the following environment variables and act
according to the spec:

- CLICOLOR != 0
  ANSI colors are supported and should be used when the program isn’t
  piped.
- CLICOLOR == 0
  Don’t output ANSI color escape codes.
- CLICOLOR_FORCE != 0
  ANSI colors should be enabled no matter what.

The behaviour regarding TERM=dumb is also retained.

This allows toggling from the global state the output of colors.

Signed-off-by: Antonin Décimo <antonin@tarides.com>