Print out more information
The command print out more information, the package in which an external
library belongs to, the libraries, executables and tests at this point
with their respective external_lib_deps.

It was possible with the command `dune external-lib-deps` which was
removed, to use `@install` `@runtest` aliasis. And we ended up with the
new command to not be able to use those, this is why knowing the package
of an external library could help.

The goal is to have much information for `opam-dune-lint`.

Signed-off-by: Alpha DIALLO <moyodiallo@gmail.com>