fix(metrics): output correct process metrics
Before this commit we would output an "async" start event when a process
awould start nd then a "complete" event when it would be finished.

The "async" start event is unnecessary and this commit removes it. All
the information recorded in the "async" start event is therefore moved
to the complete event.

The new output is now properly displayed by the various visualization
tools (perfetto, chrome)

Signed-off-by: Rudi Grinberg <me@rgrinberg.com>

<!-- ps-id: d5f448b8-a896-4612-bc4e-e8d16c41fc51 -->