fix(scheduler): flush stats when idle (#6891)
[dune --trace-file] currently relies on the output channel to flush the
events when it's internal buffer is full. This isn't a problem for a
normal build, because we only want to observe the trace file once dune
terminates, but it is a problem for watch mode. In watch mode, we have
to wait an arbitrary amount of time until the buffer gets filled up and
is flushed.

This commit flushes the events output channel in watch mode whenever
we're idling and waiting for file system events to arrive.

Signed-off-by: Rudi Grinberg <me@rgrinberg.com>