refactor: Dune_project.encode sanity
dune init relies on removing all dialects to make sure the serialized
dune-project file doesn't have any dialects. That is weird because
it explicitly relies on serializition/deserialization not being round trip.

Instead of relying on this, we change the serializer to stop outputting
the set of dialects if it's equal to the default.

Signed-off-by: Rudi Grinberg <me@rgrinberg.com>

<!-- ps-id: 75f25746-0290-4182-ac97-e1d9705969ae -->