doc: add directives to document stanzas and fields
This is about creating a sphinx domain, a plugin that knows about the
objects that dune manipulates.

This has 3 advantages:

- the documentation gets more structured and looks better
- some indexes are generated (one for stanzas and one for fields)
- it is possible to cross-reference directly to a specific stanza or field

In terms of markup, this looks like the following:

    .. dune:stanza:: mdx

       Set up rules to use with mdx. When the ``runtest`` alias is evaluated, the
       reference file will be compared to the output of the mdx tool.

       .. dune:field:: files
          :param: <globs>

          The files that are checked. Defaults to ``*.md``

This is slightly longer than before but this is more structured. It is
possible to configure a default domain so that .. stanza:: and ..
field:: are recognized instead.

This commit only deals with stanzas for `dune-project` to start.

Co-authored-by: Christine Rose <christinerose@users.noreply.github.com>
Signed-off-by: Etienne Millon <me@emillon.org>