Add support for eager watch mode for `dune exec`
This is the second attempt at adding this feature. The problem with the
first attempt was that replacing the call to `Scheduler.go` with
`Scheduler.go_with_rpc_server_and_console_status_reporting` caused
occasional non-deterministic seg faults on macos. There's some info
about the problem on the PR which reverts the previous attempt at adding
this feature: https://github.com/ocaml/dune/pull/6867. At the time of
writing we don't know the cause of this problem.

The difference this time around is that we maintain the call to
`Scheduler.go` when running `dune exec` not in watch mode, and only
invoke `Scheduler.go_with_rpc_server_and_console_status_reporting` when
running in watch mode. There is some additional refactoring done to make
this split more ergonomic. We may see seg faults on macos when running
exec in watch mode but at least we won't introduce the potential for seg
faults into the existing use case of running exec not in watch mode
(assuming of course that there is a causal link between
`go_with_rpc_server_and_console_status_reporting` and the seg fault on
macos, which is not necessarily the case).

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>