Add dune-project stanza, map_workspace_root, to control workspace mapping.
Add a "map_workspace_root" stanza to the dune-project file.
If true, references to the workspace root directory in output
files are mapped to "/workspace_root". If false, such references
are not modified.
If missing, it defaults to true.

Note that in the added tests, for some configurations quotes
are needed around the "EOF" delimeter to get expansion.

Note also that when enabled, the debug search directories in the
debug information produced by ocamlc are also mapped, with
the result that ocamldebug cannot find the files.

Fixes #6929, provided user disables mapping.

Signed-off-by: Richard L Ford <richardlford@gmail.com>
Co-authored-by: Christine Rose <christinerose@users.noreply.github.com>
Co-authored-by: Etienne Millon <etienne.millon@gmail.com>
Co-authored-by: Ali Caglayan <alizter@gmail.com>