fix: Invoke preprocessor commands from local dir
Prior to this change commands listed in the `(pps ...)` field were run
from the workspace root. This meant that the `%{project_root}` variable
couldn't be used when constructing preprocessor commands, as it returns
a path relative to the directory containing the `dune` file. After this
change the preprocessor commands are run from the directory containing
the dune file which defines them. This behaviour is feature gated as
some projects may have come to rely on the original behaviour.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>