feature(coq): composition of installed theories
This commit introduces a new database for "installed" theories present
in `user-contrib` and `COQPATH`.

The above paths are scanned and the database of theories is populated
based on the subdirectories found.

In order to keep compatibility, we add to the database all possible
theory names (so for example, given a directory `A/B` we will register
two Coq theories `A` and `A.B`).

Changes in the code are as expected, main new things are:

- `Coq_lib.t` now can represent two kind of objects: theories with an
  associated dune stanza, and theories found in the system and for
  whom no stanza is available.

- A new`Coq_path.t` type, which represents an inferred "stanza" for
  libraries that are installed in the `user-contrib` / `COQPATH`
  scheme. The list of such paths is then used to build the database.

Starting with `(coq lang 0.8)` Coq will not see any theory, even if
globally installed, unless that theory was declared in the
corresponding `theories` field in the stanzas.

This solves a longstanding problem due to users being not able to
compose with a theory that could present either locally or in
installed form.

Signed-off-by: Ali Caglayan <alizter@gmail.com>
Co-authored-by: Emilio Jesus Gallego Arias <e+git@x80.org>