adds support for loading plugins in toplevels (#6082)
* adds support for loading plugins in toplevels

Uses virtual libraries to select the proper dynamic linker
facility. To load in the toplevel add `dune-site.toplevel` library as
the dependency to your toplevel.

3. Also, handle moved load_file function.  Prior to OCaml 4.13.0, the
load_file function was in Topdirs.  Starting with OCaml 4.13.0, the
load_file function moved to Toploop. In order to find it open both
these modules, suppressing the warning for unused open, and then
reference load_file unqualified.

Signed-off-by: Richard L Ford <richardlford@gmail.com>
Signed-off-by: ivg <ivg@ieee.org>
Co-authored-by: Richard L Ford <richardlford@gmail.com>