fix(boot): fix env vars when finding out concurrency
We were passing an empty enviornment to Unix.open_process_full_cmd
before. We now pass the full available Unix.environment since we are
only querying the concurrency by calling nproc or getconf.

On NixOS this was causing the bootstrap to default to a single
processor since nproc wasn't available in the empty environment.

<!-- ps-id: d620baf8-5e58-4212-8f30-df69780a614b -->

Signed-off-by: Ali Caglayan <alizter@gmail.com>