doc: cleanup support seciton of readme and remove jbuilder mentions
It has been quite a while since the jbuilder migration so I think we
can safely remove any mention of it in the README.md.

I've also improved the support section pointing to GitHub discussions
and issues. I am not sure the ocaml-core email is being checked by
anybody so I have removed it.

Signed-off-by: Ali Caglayan <alizter@gmail.com>

<!-- ps-id: 02d7a4a4-59e1-435c-8c3f-76788a8d85dd -->