feature: introduce dune_config
[dune_config] is a library to help developers configure dune in a
consistent way. It does not aim to replace dune's existing configuration
capabilities which are meant to be stable and user facing. The type of
configuration it plans to support is experimental, undocumented, and may
be removed without notice.

For ease of use, it's possible to set the configuration values using
environment vairables and the dune-config file. the former will take
precedence over the latter.

As a demonstration, I've added a single configuration value for
disabling dune's global lock. This isn't a safe thing to do, but I know
that @snowleopard has his own locking implementation at JST and
@jchavarri needs it for a particular workflow.

Signed-off-by: Rudi Grinberg <me@rgrinberg.com>

<!-- ps-id: 9eabc13d-4e95-4bce-88ba-ff439b2df6b4 -->