add: Path.as_outside_build_dir
We already had Path.as_outside_build_dir_exn so we add the option
variant so we can decide what to do with it.

This is used in Coq's installed composition, specifically due to a use
of Fs_memo.dir_contents and adding a nice user message if _build/ was
somehow passed.

Signed-off-by: Ali Caglayan <alizter@gmail.com>

<!-- ps-id: 5a69d470-6010-4b36-8da2-a0b0cb733ede -->