doc: add todos for new structure
This PR adds todo notes to finish migrating to a new structure for our
documentation.

It is based on the [Diátaxis](https://diataxis.fr/) framework. The core
idea is to organize documentation based on what the user is after when
reading it. 4 categories emerge:

- tutorials ("can you teach me how to...")
- how-to guides ("how do I...")
- reference guide ("what is...")
- explanations ("why...")

An example of docs where this has been used is [the gatsby
docs](https://www.gatsbyjs.com/docs/).

Our existing docs have a lot of good content in it but it is difficult
to find because these concerns are mixed in the various documentation
pages.

For example, all these aspects tend to be grouped together in a single
page dedicated to a particular feature. But a reader rarely needs to
refer to every aspect of the docs for that feature.

We also have some pages ("concepts" and "advanced topics") that are
really several pages concatenated together with no strong unifying
theme.

This shows how to go to both a desired "complete" state (see
<https://diataxis.fr/how-to-use-diataxis/#complete-not-finished>).

Signed-off-by: Etienne Millon <me@emillon.org>