Support for BUILD_PATH_PREFIX_MAP to mirror the package installation.
For the testcase for issue #7413, the issue is resolved with the
current Dune changes. The result of the grep now gives a count of 1.

For the deployment phase, we need the inverse of the
BUILD_PATH_PREFIX_MAP. I've decided that this should use a different
environment variable, so I've chosen DEPLY_PATH_PREFIX_MAP.
The reason two variables are needed is that a new dune command,
`dune ocaml debug` will been added. It is similar to `dune exec`,
except instead of invoking the executable directory, it
give the executable to the OCaml debugger. But before doing that,
it builds the executable, if necessary, to make sure it is up-to-date.
During the building, BUILD_PATH_PREFIX_MAP maps build-time
absolute paths to abstract paths, but the debugger needs the inverse
mapping, so will use DEPLY_PATH_PREFIX_MAP.

So forward and inverse mappings are computed, and the
`dune describe map` command was added to display and test
the forward and reverse maps.

Promote tests with expected results for new mapping

Add find_map locally to enable build with early OCamls

Guard against invalid existing map in order to guarantee
we always emit a valid map.

Only map .ml files exactly, to decrease map size.
This is a temporary solution to avoid exceeding
the argument/environment size for execve (exhibited
in the benchmarks). Maybe make user-configurable,
or have an indirection feature in the mapping
environment variable so the full mapping can be
in a file.

Use String.is_suffix, as it is available in stdune
for all versions of OCaml.

Signed-off-by: Richard L Ford <richardlford@gmail.com>