test: Testcase for issue #7413
The test has currently been promoted to match the existing behavior,
in order not to have a failing test. When the issue is resolved the
result of the grep should give a count of 1.

Note the "touch" is after the grep, so that the overall
exit code is not affected by whether the grep matches
or not.

Do touch after grep to avoid error code

Signed-off-by: Richard L Ford <richardlford@gmail.com>