odoc: put odoc support files in odoc.support
Fixes #7364

If the directory is named `_odoc_support` with a leading underscore, it
is omitted when deploying to github pages. This is because this system
also support the jekyll site generator, which uses this convention for
data files.

We can not use any directory name here since we want to avoid conflicts
with OCaml identifiers and names of `*.mld` files.

Signed-off-by: Etienne Millon <me@emillon.org>