ci: remove MSVC 4.12.1 builds
Fixes #7592

One of these builds has been failing for some time and after some
investigation it seems that the other one is only working because of
some cached assets that can not be recreated easily.

Quoting @jonahbeckford, DkML is a rolling distribution, so fixes are not
backported to old compiler versions. At the moment of this commit, MSVC
4.14 is the current version and works as expected.

Signed-off-by: Etienne Millon <me@emillon.org>