tui: add scroll bars, help screen and debug info
The help screen can be toggled on and off.

The scroll bars appear when the main text is cut off in the horizontal
or vertical directions. They can be interacted with using the arrow
keys, j and k and using the mouse. The mouse can scroll and click on
them.

Added a utility to provide debug info, which can be useful determining
the state of the UI. It can be activated using 'd'.

Signed-off-by: Ali Caglayan <alizter@gmail.com>