tui: improve scroll handling and attribute sharing
We improve the logic for scroll handling. Notably now the arrow buttons
on the scroll bars work.

We also make many of the used style attrubutes internally configurable
and fix any styles biased to a dark background.

There are various cleanups for the drawing of the top frame, including
making the debug info menu a proper overlay rather than just appending
to the end.

Signed-off-by: Ali Caglayan <alizter@gmail.com>