refactor(engine): decouple from Dune_lang
This requires the following 3 general moves:

* Moving predicate_lang into its own library
* Moving various action definitions from Dune_lang to Dune_util
* Moving the definition of alias names to Dune_util.

The last refactoring of the aliases isn't yet completed. I think we'd
rather have rules specific aliases such as fmt and install to live in
dune_lang rather than in the engine, but for now this is good enough.

Also, this refactoring caught some missing validation bugs in the alias
module. I've added some TODO's to address them eventually.

Signed-off-by: Rudi Grinberg <me@rgrinberg.com>

<!-- ps-id: 9e878180-d0eb-4668-b5ec-1441c4301687 -->