New `dune ocaml debug` command
Add a new command, `dune ocaml debug` to set a
mapping environment variable and then invoke
the OCaml debugger, ocamldebug.

For the deployment phase, we need the inverse of the
BUILD_PATH_PREFIX_MAP. I've decided that this should use a different
environment variable, so I've chosen DEPLOY_PATH_PREFIX_MAP.
The reason two variables are needed is that the new dune command,
`dune ocaml debug` is similar to `dune exec`,
except instead of invoking the executable directory, it
gives the executable to the OCaml debugger. But before doing that,
it builds the executable, if necessary, to make sure it is up-to-date.
During the building, BUILD_PATH_PREFIX_MAP maps build-time
absolute paths to abstract paths, but the debugger needs the inverse
mapping, so will use DEPLOY_PATH_PREFIX_MAP.

This PR assumes a simplified forward mapping, namely,
the build root, source root and the install root (usually
~/.opam/switch) map to `/workspace_root`.

The inverse mapping used is (in priority order, highest
first):

- /workspace_root -> source root
- /workspace_root -> build root
- /workspace_root -> install root

Because of `dune ocaml debug`, dune requires the location of
ocamldebug. Some test cases which set up special paths had to be
modified to make sure that ocamldebug could be found.

Signed-off-by: Richard L Ford <richardlford@gmail.com>