Run short monorepo benchmarks multiple times
This is an attempt to make the benchmarks less noisy by running them
3 times and having current-bench report the mean. Fortunately the amount
of noise we see in monorepo benchmarks is constant, so long running
benchmarks (mainly the initial build from scratch) have only a small
amount of noise when expressed as a percentage of the running  time.
Thus we can get away with only running the shorter benchmarks several
times.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>