feature(pkg): Generate lockdir from opam switch
This adds an option to `dune pkg lock` for passing a switch name, which
will generate a lockfile using the given switch to solve dependencies.
This is intended as a temporary convenience so that users don't need to
download the opam repository in order to generate lockfiles. Eventually
dune will handle downloading the opam repository itself at which point
this feature will be removed.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>