add: build_if in test stanza
Fixes #6938

The semantics of `(enabled_if)` in `(test)` can be confusing:
`(test)` can be seen as the combination of `(executable)` and a `(rule
(alias runtest))`; but `(enabled_if)` actually only controls the
"running" part, not the "building" one.

This adds a new `(build_if)` field in `(test)`. When it evaluates to
false, the test stanza is bypassed (no build is attempted).

Signed-off-by: Etienne Millon <me@emillon.org>