Monorepo benchmark uses offline duniverse
This changes the monorepo benchmark to use an offline copy of its
required duniverse directory. This removes the need to run `opam
monorepo pull` while preparing the benchmark which improves the
reliability of the benchmark, as it is possible that packages can become
unavailable to download from time to time, which can cause the benchmark
to fail. This has the added benifit of speeding up the benchmark
preparation. A consequence of this change is that running the benchmark
locally requires some additional steps which are documented in the
monorepo benchmark readme.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>