Lockdirs are data-only
When a package depends on dune the lockdir will contain a file named
"dune" which is not a valid dune file (instead it contains information
about the package "dune"). To prevent dune attempting to interpret the
file as a dune file, we make all lockdirs data-only.

To simplify this (both conceptually for users and the implementation) we
introduce a convention where only directories ending in ".lock" can be
specified as lockdirs, and all directories ending in ".lock" are treated
as data-only.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>