Lockdir generation using opam switch prefers oldest
This fixes an oversight where generating lockdirs taking dependencies
from an opam switch didn't prefer the oldest version of packages, but
when taking dependencies from an opam repo in a directory we did prefer
the oldest. This changes the behaviour when using an opam switch to
match that when using a directory (both now prefer the oldest versions
of packages).

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>