refactor(engine): optimize sandbox config storage
We store a sandbox configuration set for every single action we create.
For a large build, this can add up.

Our old representation was taking 6 words to represent this. This PR
changes it to use only a single word. Moreover, this new set now opaque
to the GC, speeds up comparison, hashing, etc.

Signed-off-by: Rudi Grinberg <me@rgrinberg.com>

<!-- ps-id: 8d6ab81b-57f7-4c94-b39c-617e78e9479d -->