ci: add an action to test mirage
This gives us a known-good configuration of Mirage to test against. The
vast majority of the changes on the Dune side will not affect this, so
it is not useful to add this to the regular CI; instead, this gives us a
manual run to trigger at release time or on changes for which that can
be useful.

Signed-off-by: Etienne Millon <me@emillon.org>