Add solver env for keeping track of solver flags
The solver env is configurable per build context and contains variable
assignments used by the solver when solving dependencies. Currently it
just contains the flags with-test and with-doc but could easily be
extended to track system variables. Its purpose currently overlaps with
that of `Opam.Env.t` but the plan is to fold that into `Solver_env.t` in
a future change.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>