makefile: slightly easier bootstrapping
We introduce a phony "bootstrap" target in the makefile that will always
build the bootstrapped version of dune. Before the only way to bootstrap
properly was to use `make _boot/dune.exe` and many times this would not
work so it would have to be changed to `make -B _boot/dune.exe`.

This new target is easier to remember and write, and allows devs to
quickly rebootstrap dune when needed. This is not something that needs
to be done very often, but we make it easier anyway.

Signed-off-by: Ali Caglayan <alizter@gmail.com>