Expose user environment when running pkg actions
Currently nixos users are unable to run pkg tests that run system
actions because the `$PATH` environment variable is unset in the
environment in which the action runs resulting in the shell being unable
to find standard unix tools such as `cat`. This is intended as a quick
fix to unblock nixos users but it introduces a new problem of builds
being (unintentionally) influenced by environment variables. This will
require some careful design so we'll solve it later.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>