feat: replace deprecated runloop in fsevents
Replace the use of the deprecated `FSEventStreamScheduleWithRunLoop()`
with `FSEventStreamSetDispatchQueue()`.

Since a dispatch queue spawns new threads, we need to call
`caml_c_thread_register()` and `caml_c_thread_unregister()` for these
threads. This is done indirectly via a `pthread_key_t` so that the
functions only need to be called more than once if the dispatch queue
switches which thread is running the callback.

We replicate the blocking nature of the existing code using a mutex and
condition variable.

See git/git@b0226007f0aaf448dec1defe3e44c4e3d7513aa8 for more details
about this approach.

Signed-off-by: Kevin Ji <1146876+kevinji@users.noreply.github.com>