docs: fix mood and punctuation of flag doc strings (#8370)
* chore: stop using pre-formatted strings for flag docs

Any newlines or paragraph breaks are re-formatted away in the final
formatted manual anyway. We might as well be consistent and rely on
ocamlformat to automatically line-wrap the docs in code.

Signed-off-by: Philip White <code@trailingwhite.space>

* docs: fix mood and punctuation of flag doc strings

Fixes a few inconsistencies in flag documentation:
- Most flag docs were using imperative mode, but some were not.
- Most flag docs were punctuated as sentences (i.e. they end in a
  period), but some were not.

As I was looking over all the docs, I made a few unrelated cleanups and
re-wordings that I thought were improvements.

Signed-off-by: Philip White <code@trailingwhite.space>

---------

Signed-off-by: Philip White <code@trailingwhite.space>
Co-authored-by: Philip White <code@trailingwhite.space>
Signed-off-by: Diana Kalinichenko <dkalinichenko@janestreet.com>